Implemented frost beam bullet functionality

+++ Space-Invaders/source/Bullet/BulletController.cpp
@@ -96,32 +96,41 @@ namespace Bullet
 
 	void BulletController::onCollision(ICollider* other_collider)
 	{
-		PlayerController* player_controller = dynamic_cast<PlayerController*>(other_collider);
-		if (player_controller && getOwnerEntityType() != EntityType::PLAYER)
-		{
-			ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
-			return;
-		}
+		processPlayerCollision(other_collider);
+		processEnemyCollision(other_collider);
+		processBunkerCollision(other_collider);
+		processBulletCollision(other_collider);
+	}
 
+	void BulletController::processBulletCollision(ICollider* other_collider)
+	{
+		BulletController* bullet_controller = dynamic_cast<BulletController*>(other_collider);
+		if (bullet_controller) ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
+	}
+
+	void BulletController::processEnemyCollision(ICollider* other_collider)
+	{
 		EnemyController* enemy_controller = dynamic_cast<EnemyController*>(other_collider);
+
 		if (enemy_controller && getOwnerEntityType() != EntityType::ENEMY)
 		{
 			ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
-			return;
 		}
+	}
 
+	void BulletController::processPlayerCollision(ICollider* other_collider)
+	{
+		PlayerController* player_controller = dynamic_cast<PlayerController*>(other_collider);
+
+		if (player_controller && getOwnerEntityType() != EntityType::PLAYER)
+		{
+			ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
+		}
+	}
+
+	void BulletController::processBunkerCollision(ICollider* other_collider)
+	{
 		BunkerController* bunker_controller = dynamic_cast<BunkerController*>(other_collider);
-		if (bunker_controller)
-		{
-			ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
-			return;
-		}
-
-		BulletController* bullet_controller = dynamic_cast<BulletController*>(other_collider);
-		if (bullet_controller)
-		{
-			ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
-			return;
-		}
+		if (bunker_controller) ServiceLocator::getInstance()->getBulletService()->destroyBullet(this);
 	}
 }
